requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.6.1
MarkupSafe==3.0.2
msgspec==0.18.6
node==1.2.2
npm==0.1.1
odict==1.9.0
openai==1.52.1
optional-django==0.1.0
packaging==24.2
pluggy==1.5.0
plumber==1.7
pydantic==2.9.2
pydantic_core==2.23.4
pytest==7.4.2
python-dotenv==1.0.1
requests==2.26.0
setuptools==75.3.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
Werkzeug==3.0.4
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==7.0
zope.interface==7.1.1
zope.lifecycleevent==5.0
zope.proxy==6.1
